implementing Tower class and adding some changes to Missile class.
I dont know basic informations about timerTick and its fuctionality.
I didn't implement that.

Changelog:

in Tower:
changing the output of getNearByenemies to Cell
changing the name of getNearByEnemies to priorAttakingEnemy
changing the name of getCannonPower to getMissilePower
reomving getCell method because the parent already have it

in Missile:
adding setter and getter methods
removing setting target in constructor because it changes over time

in Cell:
adding the default constructor